football_web/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
python-multipart==0.0.9
slowapi==0.1.9
pyotp==2.9.0
qrcode[pil]==7.4.2
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
aiosqlite==0.20.0
